+++ services/server/src/server/apiv1/repository/get-contract-addresses-all.stateless.paths.yaml
