+++ services/server/src/server/apiv1/repository/get-file-static.stateless.paths.yaml
