+++ services/server/src/server/apiv1/repository/get-source-files-full.stateless.paths.yaml
