+++ services/server/src/server/apiv1/verification/vyper/stateless/vyper.stateless.paths.yaml
